Adding hidden imports
This are imports that generated files will use. These are not needed for
correctness, but when the generated source files are not in the source
tree, it makes a few of the tools behave a little better.

Also adding doc.go to packages that did not have one and moving the
package comments into them.