Vulkan:Emit SIGTRAP in memory tracker before passing SIGSEGV to underlying handler
In Debug build, if a non-tracking-coherent memory triggers SIGSEGV
signal, raise the SIGTRAP signal for debugger to capture, before passing
the SIGSEGV signal to the underlying segfault handler.

In this way, when debugging tracing with memory tracker enabled, even if
the debugger is set to skip SIGSEGV signals (since it is triggered too
oftenly), it can still capture the invalid memory accesses.