Handle our own UI-thread queuing for a snappier UI.
It is common for multiple updates to be queued from the RPC threads in
quick succession. This caused some 2+ seconds delays as the UI thread
seems to only process one queued item at a time. Instead, batch up all
queued request and execute them together as soon as the UI thread gives
us control to do so.