Fix an NPE and be more specific about default settings.
On first launch, or when the settings file is not found, the load
is skipped, which caused an NPE due to unintialized fields. This
change moves all the default settings out of the updateFrom method
to the field defintions, instead of having them split across the two,
and is more explicit about default values that were implied.