Replace minRequiredVersion with @if
The intention of @if is the same of as of #if. If the condition
is not true, the semantics is as if the code block was deleted.

This allows us to express more complex conditions which were not
possible before (e.g. Version.GLES30 || Extension.GL_FEATURE).

It also fixes the problem that checks were not possible in switch
expressions, and resulted in incorrect error in switch statements.

Extending this idea to fields will allow us to filter irrelevant
data from the UI.