Add StartOfFrame annotation (#532)
* Add BeginOfFrame annotation

The new BeginOfFrame is exactly the same as the original EndOfFrame
annotation (as shown in resolve/events.go)
For GLES, atoms like eglSwapBuffer will be annotated with BeginOfFrame,
as the framebuffer image should be read before mutating such atoms. This
CL should not change any behavior on GLES side.

For Vulkan, vkQueuePresentKHR will be annotated with the new EndOfFrame
annotation. So when grabbing framebuffer images, vkQueuePresentKHR will
be mutated (but not called on replay side), so the server will grab the
image which is to be presented, not the image just rendered.

* Rename BeginOfFrame to StartOfFrame, revert some incorrect changes.

* Revert another touched java code