Vulkan:Postpone the messaging of framebuffer observation atom (#533)
Gather the framebuffer observation atom data prior to the EndOfFrame atom:
vkQueuePresentKHR, but do not encode and message the atom. The atom will
be messaged after the vkQueuePresentKHR, so that on the replay side, let
the read_framebuffer.go grab the image to be presented first, then we
can compare the framebuffer images from trace and replay sides for the same
vkQueuePresentKHR atom.