gapic: Workaround late cancelation in SingleInFlight.
By the time SingleInFlight cancels the previous future, it is too late
and so both the old and new futures are actually cancelled. The "single
in flight" logic needs to happen before the futures are created, so the
second future is created only after the first is cancelled. This needs
some extra work and will happen soon(ish). For now, just work around the
issue, which was uncovered due to the "double searching".