Robot: Adds packaging of build files to do upload.
Implements an auto packager for the upload verb for robot. First we look to see if the package file
already exists, if not we create a zip file, populate it with the build artifacts in the directory
structure expected by test/robot/build/artifact.go (TODO(baldwinn): currently hardcoded, either
expose artifacts desired layout as a FileLayout, or make it use the FileLayouts defined in
core/app/layout/layout.go).

FileLayouts from core/app/layout/pkgdata.go are used to find the files to be packaged, and a method
to retrieve the path to gapis was added.

Changed upload to respond to errors from uploader.prepare().

This also adds .dylib handling to the layout package so it naturally works on darwin (though not
entirely necessary, this makes more sense than adding special code to change the path in the robot
command to .dylib on OS X as a side effect of [this bug](https://github.com/google/gapid/issues/570).
Fixes not passing through ctx on layout.Gapis().