Robot: Change ToolSet, add build based artifacts.
The decision was made to tie android tools directly to the host that built them within an artifact. This allows Kokoro to simply upload an artifact at the end of each host's build step. Another potential problem this solves is when a bundle
may have failing android tools that had never been tested because merging logic caused the master to ignore them.

Now a toolset is a collection of a single abi and hosttoolset from the build device and an androidtoolsets for each
android ABI that the build device created. Now when searching for a target androidtoolset when tracing, a host abi must
be given at the very least.