Fix file mode for zipped tools, Android abi list.
Zipped tools were losing their execution bits because the created file in the zip writer was defaulting to read-write
but no execute permissions. Creating the file with a file header fixes this.

Android ABIs were being tested just based on architecture, which cause binlayout packages to fail since they need the
abi name to find the directory. Now we list the supported device ABIs directly (though this should be moved to a better
place still).