Fix stash and shelf URLs on Windows.
Creating stash and shelf files on the master server involves performing
a URL resolution to find the correct protocol scheme and location of the
file. The default method is to just pass the file path and assume the
resolution will figure it out, unfortunately on Windows, the URL
resolution will see the volume label followed by a colon and take that
to be the protocol scheme, which causes the server to fail.

The fix is to generate a correct file:/// url from the file.Path and use
that instead of the plain filepath, and then remove the extraneous
forward slash that the URL processing code prepends to the location
field of the URL on Windows.