Robot: Webserver works on Windows.
The embed command was using system dependant filepaths for use in
lookup. This was an issue in static webserver which attempted to lookup
files based on the request URL, which uses different filepath separators
from Windows. The fix is to add an option ('WEB') to embed that swaps
filepath separators with url standard forward slashes.

There is also an issue when loading the root index, which used
filepath.Base to look for a forward slash which will not be the case in
Windows. The fix is to use filepath.Separator to look for a root name
instead.