Robot works on Windows!
There were several problems with filepath mismatches because some calls
expect tools to be in binLayout order, or at least have that naming
convention. Switched the upload tool and artifact processor to use that
layout instead now.

ffmpeg and git could not be found because we clear the path out and
don't try to add the files they are in!

Windows has ".exe"s and bin layout didn't care before, now it does.

There needed to be a way to get the android abi name from the binlayout
directory as well.

Many instances of errors never being returned in error states because it
halts the scheduler, so I switched it to storing a slice of errors and
just returning and printing them out once done.