Add histogram data to report.
Gapit report will now optionally output histogram data with the flag
'-histogram'. This data is in the following format:

[Histogram] frame #{1}, [{2},({3},{4},{5},{6})], ... )]
1 - the number of the captured frame that was tested
2 - the difference bin* of this group.
3 - the number of pixels with red channels in this bin.
4-6 - the number of pixels with green, blue, and alpha channels in 2
* each bin represents a difference range to calculate the range do the
following:
min := (bin-1)*255/(binMax+1)
max := bin*255/(binMax+1)
binMax is the number of bins total (currently 32)

This is a quick and dirty way to get the histogram data over to robot.
Better representations will most likely lead to better performance and
usefulness.