Fix object view in robot webserver and uploads.
When the format of build ToolSets changed it broke a part of the data
gathering functionality in the webserver. There was already a TODO to
change it to use the proto directly rather than using the json and
assuming the format doesn't change so I took it as mine for the future.

The master server was only listening from localhost for uploading
subjects and artifacts. Now it accepts from any valid connection, so I
can upload from another computer.