gapis/gles: Fix FB requests stalling forever.
If you request a frame at a command that aborts, the DCE will remove that command from the execution list. The `readFramebuffer` transform then never sees the CmdID it's waiting for, and the request is never completed.

Instead, change `readFramebuffer` to keep an ordered list of requests. If the next command is skipped, read the framebuffer at the next command instead.