Fix "text file busy" in robot.
An unfortunate side effect of unix' fork/exec model is that between fork
and exec file descriptors marked for close-on-exec are still open in
the child process. This is usually perfectly acceptable, but not when
writing and executing test binaries in quick succession on separate
threads. This fix is very ugly, but is currently the accepted solution
to this problem, more details about it are in the large hack comment in
the file.

Also to see how many processes this fixes, added a log message to Wait
when there are no errors.