Reverts error handling for Trace client.
Turns out, the way client error handling was done previously was
correct. The command's error needs to be returned at the end of the
function, after uploading logs and the trace file. This was making
failing traces no longer upload their logs, so we couldn't figure out
what was wrong with them.