Add build dimensions to robot workers.
The robot webserver needed to have a better understanding of packages
and builds. Prior to this change builds were typically looked at through
the lense of the GapidAPK and Gapit versions that were used to trace.
This caused multiple issues aside from just being not user-friendly.

We now have a package id used with the input for each worker, and the
webserver recognizes that as a separate dimension. A separate dimension
is used to hold the build that was used for tracing specifically, since
that feeds into the input for Report and Replay. The Host dimension now
identifies the host machine that runs the worker, rather than the host
machine that ran the trace which is less useful, as we already have the
target as a dimension.

Also need a template function mapping to test if a package type is User
or not.