Only run workers on traces with matching packages.
It is not scalable to run report and replay on every historical trace
every time a new package shows up. Eventually Golden Traces will be used
to run report and replay no matter the package, but for now lock the
workers to only use the same package for all actions. Also remove the
tracePackage dimension.