Add parent concept to task representation.
In order to properly show the change in result from task to task there
needs to be a form of history tied to the task representation. To
achieve this we now track the parentage of each package as we enumerate
them and when getting tasks we use the package parentage to create a
task parentage. Currently any task not marked as Current will also
update its result to match its parent's result.