Add regression and fixed status to tasks.
We want to be able to, at a glance, see the positive or negative changes
a particular package has, this change allows us to highlight the fixes
(successful tasks with failing parents) and regressions (failing tasks
with successful parents). Currently regressions are colored a dark red
while fixes are colored a light green.