gapii: Pass the path of libinterceptor down in the header.
This moves the loading of libinteceptor.so from jdwp_loader.go (gapit) to gapii.

It appears that on some devices dlopen() will not find libraries loaded by a prior call to System.loadLibrary(). Instead, pass the absolute path of libinteceptor.so down to libgapii.so and load it directly there.

Fixes: #1026