Fix scheduler issue.
This was a very subtle bug which caused robot to not schedule any
reports or replays after a package was added to a track. Package update
is called after it is added to a track in order to propogate the parent
information. But it is called with only a partial package, expecting the
old package to merge (which it does) but then the event is triggered
after apply using the new package (which is still only partial) which
causes all listeners to recieve (and potentially replace) copies of the
package which do not contain any tools.

Essentially packages in the scheduler were having their tools wiped out
whenever they were updated by the track, which meant we could no longer
replay or report. Traces only worked because they were scheduled after
the package's first event trigger, when they still had tools.