Replace builder abi, allow BuildBot when git present.
In the case that an abi is provided from the command line we will use it
instead of the host uploader's instance information. This is to allow
prebuilt packages to be uploaded by another computer, esp. in the case
of BuildBots.

Submitting a BuildBot build when git is present would still change the
type to User. Now the test is if a CL is specified in the command line
it keeps the BuildBot type, if prepare needs to ask for the head CL,
then it is a User build.