Add auto constraints selector.
Certain dimensions need the ability to automatically select the
constraints based on the state of the page. This change adds the auto
selector which calls an overridable function that can manipulate the
constraints variable from the dimension itself. For now this means that
the trackDimension will constrain itself to the master or auto track and
the packageDimension will constrain itself to the latest package of
current track unless there are too many other constraints, in which case
it acts like the all selector (a rename that was previously any).