Use 'sint' for most proto fields (breaks file format)
This will avoid breaking compatibility in the future
if we change, say GLint to GLuint, somewhere.

'sint' can represent all values including unsigned.
For example, 0xFFFFFFFF can be compactly stored by
by cast to int32 (-1) which is stored in one byte.

All command arguments are in the app pool by definition.
Remove the pool ID as it is redundant and just takes space.

Renumber keys in commands.