Don't use virtual trees, they are slow.
Virtual trees appear to have some overhead that makes scrolling slow
on OSX. Non-virtual trees on OSX also have bad overhead when
changing the icon, so this change also uses a custom painter for
the tree labels (only the labels, not the tree itself).

Refactored (most of) the duplicated code between the command and
state views into a re-usable tree widget.

Fixes #384
Fixes #683, since we now the image is part of the label
May be a work around for #941