Allow do upload to specify flags to robot, branch->track.
In order to allow `do upload` to forward command line options to the
underlying `do robot` commands they must be represented in the original
verb or else `do` will cause an error for unknown flags. This change
factors out those flags into UploadOptions and RobotOptions and replaces
the necessary verbs on `do` and `do robot`.

This also changes the cli flag of branch to track as that is the true
meaning of the flag, it just uses the branch when a user build does not
specify a track manually.