Add framework for anonymous user statistics (entirely disabled)
This adds the foundation code for reporting anonymous user statistics using Google Analytics.

This CL does not enable any sort of tracking - the code is completely disabled and removed from the build. In order to be turned on, another CL needs to uncomment the code annotated with `TODO: <ANALYTICS>` and the user has to agree to sending anonymous user statistics to Google.

I’d like to get this submitted sooner rather than later so I don’t have to continually rebase this branch.