Fix for multiple devices in one system.
Some systems, especially laptops, may have multiple adapters that
both support Vulkan. Furthermore, some drivers returns these
adapters in an effectively random order, or at least not
guaranteed to be the same between the tracing program, and
gapir. This remaps the physical devices on replay, so that
commands that were directed to a specific physical device are
directed to the same physical device if possible.