Add entity previews to object view.
In order to preview entities without needing to leave the grid this
change adds a video dom element and modifies the object view to create
previews of logs, reports, and videos in the grid view. Because this
makes the object view larger in size it also made sense to move it to
the right side of the screen and position it absolutely (so you don't
need to scroll to the top of the screen either).