Add stale unknown state to grid.
Previously any task that was added to the grid, but yet to have started
would have no status associated and would get labeled as Stale/Failed.
This caused it to look similar to a failure case. Instead it is now
labeled as Stale/Unknown as it is not in progress yet, but it has been
scheduled thus the status is not yet known. It will show up in the grid
as grey instead of red.

This also fixes a bug where the template lookups for target and host
devices had missing braces, causing those templates to fail to match.