core/fault/stacktrace: Implement compression.
The goal here is to get a stacktrace down to something under a 150 bytes so it could be reported in a Google Analytics exception message.

I've implemented a bunch of different compression methods, compared these to flate, gzip, lzw and zlib. All of the custom compressors out-compress the standard compressors.

As compression payloads are small, and time isn't that important, the algorithm tries to compress with 8 different compressors - and encodes the compression type as the first 3 bits of the data.

Currently unused, but hopefully will be soon.