Show a warning if the image has an empty alpha channel.
If the image has an alpha channel, where all values are below a threshold,
a warning and option to disable the alpha channel are shown. If the alpha
channel is disabled through this option and later an image with an alpha
channel is selected, a warning and option to re-enable the alpha channel
are shown.

Fixes #723