core/app: Use an redirectable log.Handler.
The old GAPIS logging system would replace the log handler on the context with a log.Broadcaster.

This didn’t work as the app framework would close it’s handler on app termination, but due to context immutability, these close messages wouldn’t propagate down to the newer broadcaster. This causes messages to not get flushed, and some logs were missing tail messages.

Fixes: #1113