Fix monitor/scheduler race condition.
There was a race condition that caused the scheduler to update robot
tasks with incomplete information. Managers would monitor for changes
inside a goroutine that would compete with the update logic of the data
owner. All initial data would be read in at once, but still in another
thread, which caused the scheduler to update without full knowledge of
entity state, this would cause extraneous tasks to get spawned every
time robot was restarted.

The fix is to initialize the owner's data with a non monitoring search
in the main thread before spawning goroutines to continue monitoring
updates. This ensures we have the complete initial information prior to
the first update.