Fixed replay of partial command-buffers that contain pushconstants.
We were getting the data for the U8s, which was not correct.
Instead read the data out into a []uint8, and then upload
that in the recreated command. This has the desired effect
of pushing the right data into the push constants.

This also fixes the same problem with VkCmdUpdateBuffer.