Override the OSX scrollbar behavior for the image panel.
Forces the use of the "always visible" style of the scrollbars on the
image panel, instead of the possible overlay style, which would hide
portions of the image.

Introduces a platform specific SwtUtil class inside the the SWT package,
so it can easily access protected fields, which can be (sparingly) used
to customize platform specific behavior.

Fixes #1141