Adds breakpad integration in GAPIR.
Breakpad is a library for capturing crashes and constructing minidumps
in native code across Windows, Linux, OSX, iOS, and Android.

The CrashHandler class registers an exception handler and adds a callback
function to perform any extra operations after the minidump has been
created. A flag is added to GAPIR to enable crash handling and reporting.

The CrashHandler class made most sense to add to core/cc since it is not
intrinsically tied to GAPIR, and may extend into GAPII at some point.

Another submodule for linux syscall support library in order to build the
Android/Linux static libs for breakpad is also added.