Fixes false positives in sxs video for empty frames.
If a frame was empty it would ignore the difference it had with the
frame buffer observation and not report an error in the side by side
video generator. This makes a bunch of replays on robot look green when
they are anything but. To fix this I just removed the check for
undefined video frames before the error report, I'm not even sure why
undefined frames were considered acceptable in the first place.