Fix trace failures not uploading logs.
Despite this being against golangs ideomatic style to returning errors
only with nil objects, the clients return "valid" output objects during
an error, this is to communicate logs and error strings that otherwise
could not get uploaded. A fix changed this behavior in the trace client,
and it has not uploaded any logs since.

This fix reverts that change since the function is unexported, and the
alternative would be to populate an output parameter, which also goes
against golang style.