Fix side by side video replaying wrong command.
The sxs video replays were specifying the incorrect command to capture
for replay and thus were retrieving an undefined replay frame. This is
effectively a revert from a previous change. The reasoning is that the
observation itself will show up as a command index and to get the
correct end of frame command we must subtract 1 in order to skip the
observation.