gapit: Fixes for SxS videos.
FramebufferObservations can happen on start of frames and end of frames.
Examples are: `eglSwapBuffers` and `gvr_frame_submit`.

For `gvr_frame_submit`, this was working correctly - when you request the frame for the framebuffer observation, you asked for the frame after the `gvr_frame_submit` command. The gvr library has some magic to correctly hand back the frame associated with the submit call, as the currently bound framebuffer could be anything.

For `eglSwapBuffers`, things were broken. Asking for the frame after a swap buffer correctly gives you back the undefined framebuffer pattern (as the contents is not preserved between swaps).

To fix, keep track of writes to the framebuffer. When you see an observation, compare this to the command that last wrote to the framebuffer.

Tested on GVR and GLES traces, both SxS and regular.