GAPIC (native) launcher script.
Adds a new binary "gapid" that will be part of the distributed package.
The binary looks for the JVM (bundled or from the system), the GAPIC
JAR (bundled or from the build output) and launches GAPIC with the
correct JVM flags and environment variables.

The launcher adds the following flags:
--vm <path to JVM>: Overrides the JVM search and runs GAPIC with the
                    given JVM.
--vmarg <arg>: Adds the given arg as an argument to the JVM. This flag
               can be repeated and each passed argument will be added.
--console: (Windows) Causes the launcher to allocate a terminal console.

The launcher script also intercepts the --help flag and appends its
help information after GAPIC's.

The main reason for this launcher is so that we can launch GAPIC on
Windows without creating a terminal (the binary is marked as a GUI
app and it launches javaw, unless --console is given).