deviceinfo: Do the device instance querying on a separate thread.
The queries for GL information was corrupting the currently bound EGL context. Instead of jumping through hoops to store and restore the context, just do the query on a separate throw-away thread.

Also clean up the messy logic in `Spy::get()` which was unnecessarily complicated and can now be simplified due to the explicit installation of the interceptor.

Fixes some traces where contexts were corrupted when traced.