Handle the OpenDocument event in GAPIC.
This event is sent on OSX if the user requests GAPID to open a gfxtrace
file, for example via double-clicking on a gfxtrace file in the Finder.
This event is sent if the application is already running, or if it just
has been started in response to opening a file.