Kill findOrZero.
Just use the normal map indexing operator.
This mean indexing might modify the map.

This fixes a SharedMap bug which is caused by the fact that
findOrZero allocates one static global mapped type to return
in the cases where the key was not found. This ultimately
results in this global value being potentially shared.