Workaround for apparant JFace bug with selection preserving.
The stack trace from #1264 indicates that there is a bug in JFace's
selection preserving when changing the viewer's input. The way our
viewers are setup, only the root node is in the loaded state when the
viewer's input is changed. JFace, however seems to query children of
a yet unloaded node, which causes the failed precondition.

This changes clears the selection just before changing the input,
which will cause the selection preservation to be skipped. This will
not only work around this issue, but also improve performance, since
right after setting the input we also fix the selection anyways.

Fixes #1264