Fix findOrZero.
Return the entry by-value rather than by-reference, same as go.

If the key want not found, the by-reference version returned
reference to single global statically allocated instance,
which eventually led to multiple consumers sharing the data.
The result was returned as const but didn't completely
protect it - we ended sharing references to maps.