Updates to the generated GAPIC version file.
Makes the version class a proper Java file, so it can easily be extended
and adds a new GapidVersion class which just contains the constant of the
current version, which is the only thing that is actually generated by
cmake on build. Minimizes the amount of Java code that is generated.

Also moves the .in file to the proper package location, rather than the
gapic/ root.