Adds crash handling and reporting to GAPIC.
This change adds an UncaughtExceptionHandler implementation that uploads a Throwable's
StackTrace to Crash2 for processing, as well as a static report function to manually
upload a StackTrace from another handling context. Currently this targets the staging
report area which only accepts crashes from internal IP addresses and is completely
disabled until the UI to enable crash reporting is created.