Make map!(K,V) a reference type in C++ to match Go.
Match the semantic of the two languages for sanity.

I choose reference type over value-type as I have found
strong need for it while writing code in the api language.

It also changes findOrZero to return result by-value, same as go.
If the key want not found, the by-reference version returned
reference to single global statically allocated instance,
which eventually led to multiple consumers sharing the data.
The result was returned as const but didn't completely
protect it - we ended sharing references to maps.