Remove resource IDs on-disk. Reference them by index instead.
The first saved resource has index 1, the next has index 2, etc...
Memory observations and others can use the index to reference it.
Protos use the index for referencing, but go still uses the hash.

This saves over 10% of the trace data (excluding raw resource bytes).

Also: Remove Pointer message as it is no longer used, and
reserve Observation.Pool field - I expect it to be useful for
storing memory observations in the initial state block.