Fixes an issue with the filmstrip when opening multiple captures.
The timeline model did not reset itself when opening a new capture, but
delayed the reset until the contexts were loaded. This caused the
filmstrip to be in a bad state in the time between the new capture being
loaded and the contexts of the new capture being loaded. At best this
caused the old thumbnails to re-appear and at worst an error dialog, if
the selection would be out of bounds in the new capture.