Limit the number of items in the texture reference popup.
This is a work around for now, that I think is good enough, because I
think the main use case for this dropdown is to find where the texture
was initialized or track down small usages of unexepcted textures. The
long term solution would be to create a custom widget that can handle
displaying a lot of references. This is tracked in #1332.

Fixes #1327