Resolve function pointers for extensions we do not support.
This means that if we come across an unknown function pointer,
it is less likely to crash.

This does run into the problem that the state block may
end up in an inconsistent state because functions are not
tracked. However without this, any calls to the functions
will just be a segfault, and even more difficult to track down.