core/app: Log crashes to the logger.
The panic message was not being displayed in GAPIR as GAPIS now correctly cleans up on exit (crash or normal). Part of the clean up is closing gRPC connections and tearing down the server. GAPIT exits before GAPIS if the connection is aborted.

This change let's any loggers or remote listeners be informed of the crash before the connections are torn down.

Fixes: #1365