Adds external crash reporting to GAPIS.
In order to accept minidumps and external crash reports in GAPIS it is
necessary to change the crash reporting functionality to accept a more
general idea of a "Report" rather than a generic interface and
stacktrace. Specifically the idea of a filefield needs to be used in
order to properly upload minidumps in the format expected by the crash
API.

To get information parity with external reports, the replay
executer also needs to get the OS of the target device rather than using
the OS that gapis is hosted on. Several other changes to tests and
reporter functions needed to happen as well to match the new API.

The shutdown CancelFunction also needed to be moved out of the atExit
reporter as that was causing the whole app to close when that reporter
was called, often preempting the uploading reporter. As well it caused
external crash reports to close immediately close the app, which may not
be desirable.