Adds crash reporting functionality to GAPIR.
We chose to upload crash dumps from gapir to gapis via the existing
server connection. This change adds a file reader to access the contents
of the minidump file as well as a crash uploader that handles piping the
crash handler function into the server connection while grabbing the
minidump contents.

Other changes were made to use the temporary directory for storing the
minidumps on platforms that support one (i.e. not Android), and allowing
the handler function of a crash handler to be set and unset (going to a
default handler).