Build GAPIR with MSVC on Windows.
This change adds the ability to build GAPIR with msvc on Windows in
order to produce pdb symbols. Most changes have to do with adding an
external project that specifies msvc as the compiler, as well as finding
the correct libraries and setting the correct compiler flags when
running under that project.