WIP: GAPIR Symbols on Windows.
This is a big mess at the moment full of in progress debug messages and
temp code. The plan is to build GAPIR with MSVC on windows so we get a
pdb file that we can use to dump the symbols. This has proved a very
tricky task.

The other part of the task is moving the symbol dumping into a cmake
script which was just started in cmake/Sumbols.cmake.