Output GAPIR symbols on Windows.
This change fixes a build error with msvc where a log line could not be
used inside a lambda under the core namespace. Several msvc specific
compile flags were being ignored as well.

Configurations that specify MSVCWinGapir now will create a single
linktime generated pdb file that then is processed with dump_syms to
produce a symbol file that can be uploaded to the crash server.