Skip contexts that were never bound in the replay cleanup.
At the end of a replay, the contexts are cleaned up by deleting all
resources. This change will cause the cleanup to be skipped for contexts
that were never bound. If an application creates a context, shares it
with another, but never binds it, we would try to clean up the resources
in the never-bound context. This would cause a bunch of errors in the
replay, since the backbuffer for this context was never initialized due
to the fact that this is done using the extras from any observed
eglMakeCurrent, which are non-existent for any never-bound contexts.

Fixes #1435