gles: Fix for cross-context deletions.
It is a no-op to delete objects that do not exist.
GAPID uses a shared context for replay - while its fine to delete IDs that do not exist in the current context, they might exist in another. Strip out any IDs that do not belong to the current context in the compat layer.

Fixes: #1446