Update adb devices in the registry when their status changes.
If a device's status changes in adb, this change will cause it to be
removed from the registry and re-added, in order to keep the device
configuration information up-to-date. When a device is added to the
registry, it's configuration is queried, which fails if the device is not
online and the device is added with empty/incomplete information. With
this change, such a device is updated once it becomes online. This is
especially needed for devices needing user authorization, since we're
likely to see the un-authorized device long before the user can give
permissions in the dialog.

Fixes #1445