Fix object deletion compatability.
In order to check whether a glDelete* command is deleting an existing
resource, the compat code reads the array of resource ids of the delete
command. However, it is currently doing so without first applying the
observation to the memory pool and thus does not actually use the
correct values.