gles: Fixes for unused VAAs that have no data.
On some drivers having a vertex attribute array enabled with no bound data will cause `GL_INVALID_OPERATION` errors, even though that vertex attribute is not used by the program.

This is a compat workaround for this case where we temporarily disable the vertex attribute array if we believe the program does not use it.