Add VulkanDriver to the device::Instance trace header (#1501)
An empty VulkanDriver will be created along with GL driver and other
generic information if Vulkan loader can be found. Real content of
VulkanDriver will be populated in the first vkCreateInstance call of the
tracing application.

When VulkanDriver is populated, another header with updated
device::Instance will be sent (ID rehased also). Decoder of the trace
file should always use the latest header.

States of VulkanDriver in the header:
1) Does not exist: The tracing device does not support Vulkan.
2) Exist but empty: The trace does not contain vkCreateInstance, the
trace should be replayable on all ABI compatible Vulkan device.
3) Exist and not empty: The trace contains vkCreateInstance, replay will
be initiated only on same vendor/device/api version device.

Also: use the -gapir-device to select the device to run gapir