Fix the gapidapk crash without force-kill.
Force-kill won't work for robottester, as `gapid devices` can be
queried during a replay of one of the testing devices, force-kill will
stop the replay and make the corresponding host hang.

Intead, the real problem is that we missed the event looper in the
native activity. Adding the looper back fixes the problem.