Added a TempSlice mechanism that does not copy.
We hit a bad edge-case where a ton of strings were being used,
We were allocating and freeing these new ranges a lot, and they all
ended up being large. Instead, return a temporary slice into the
original range. During the StringSlice, there is no opportunity to
modify the pool, so we can save the pressure on the GC.