Add upload option for subjects with OBBs.
Passing an obb file via the `-obb` flag to `do robot subject upload`
will now upload the file and append it's id to the subject specified.
This does not work with multiple subjects as there isn't a clear way to
associate obb files with their subjects, so this is explicitly
disallowed.