Removes OBBs after tracing, grants R/W permission to APK.
To remove the OBB when finished we add a deferred call to clean it up.
Also the APK may not have specified read and write permissions for
external storage as it seems the Unreal install scripts typically add
this as an extra step, so we grant it as well, only if there is an OBB
though.