Switch robot web client to MVC model, add defVals.
This change is largely refactoring and juggling code into a framework
that can be ported more simply over to javascript. Following a MVC model
the page struct and some other methods have been moved and split between
a view and controller struct that separate the concerns. The controller
handles setting values in response to input, the view handles displaying
those values in the selecters and the grid.

defVal() is added to the dimension struct in order to have some method
of defaulting values when first entering the page.